split graft recovery helpers below source line cap
Finding: AW4-QA-B1
